chore: moved counter and manual data binding example to playground folder
The counter example has been moved out of App.js and into the playground folder.
Rendering has also been moved back to index.js.
