+++ src/index.js
@@ -9,7 +9,9 @@ var template = (
     <App />
   </React.StrictMode>
 )
+var appRoot = document.getElementById('root')
 
+ReactDOM.render(template, appRoot);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
